pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "velox"
version = "0.1.0"
description = "High-performance, async-first ETL connector library with multi-environment support"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [
    { name = "Velox Team" }
]
keywords = ["velox", "etl", "kafka", "async", "connectors", "data-pipeline"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Framework :: AsyncIO",
    "Typing :: Typed",
]

dependencies = [
    # Async Kafka client
    "aiokafka>=0.10.0",
    
    # Fast JSON serialization (Rust-based)
    "orjson>=3.9.0",
    
    # Configuration and validation (Rust core)
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    
    # Template rendering
    "jinja2>=3.1.0",
    
    # Environment variables
    "python-dotenv>=1.0.0",
    
    # YAML parsing
    "pyyaml>=6.0.0",
    
    # Structured logging
    "structlog>=24.1.0",
    
    # Rich CLI output and progress
    "rich>=13.7.0",
    
    # Async utilities
    "anyio>=4.2.0",
]

[project.optional-dependencies]
# Fast event loop (Unix only)
uvloop = ["uvloop>=0.19.0; sys_platform != 'win32'"]

# Development dependencies
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "mypy>=1.8.0",
    "ruff>=0.1.0",
    "black>=24.1.0",
    "pre-commit>=3.6.0",
    "types-PyYAML>=6.0.0",
]

# All optional dependencies
all = [
    "velox[uvloop,dev]",
]

[project.urls]
Homepage = "https://github.com/velox-connectors/velox"
Documentation = "https://github.com/velox-connectors/velox#readme"
Repository = "https://github.com/velox-connectors/velox"

[project.scripts]
velox-run = "velox.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/velox"]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/config",
    "/templates",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
exclude = ["tests"]

[tool.ruff]
target-version = "py311"
line-length = 100
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "ARG", # flake8-unused-arguments
    "SIM", # flake8-simplify
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.isort]
known-first-party = ["velox"]

[tool.black]
line-length = 100
target-version = ["py311"]

